RECT int instead of long

+++ RFReborn.Windows/Native/Structs/RECT.cs
@@ -11,21 +11,21 @@ namespace RFReborn.Windows.Native.Structs
         /// <summary>
         /// Specifies the x-coordinate of the upper-left corner of the rectangle.
         /// </summary>
-        public long left;
+        public int left;
 
         /// <summary>
         /// Specifies the y-coordinate of the upper-left corner of the rectangle.
         /// </summary>
-        public long top;
+        public int top;
 
         /// <summary>
         /// Specifies the x-coordinate of the lower-right corner of the rectangle.
         /// </summary>
-        public long right;
+        public int right;
 
         /// <summary>
         /// Specifies the y-coordinate of the lower-right corner of the rectangle.
         /// </summary>
-        public long bottom;
+        public int bottom;
     }
 }
